Fix wrapping connection with dynamic parameters
When a connection has parameters defined dynamically at runtime, the
same wrapped connection was used because the cache key was only based
on the connection's name, leading to using a connection that does not
honor the parameters.

Always using a cache key based on all parameters (including connection's
name) prevents this.